requirements.txt
ccxt>=4.0.0
pandas>=2.2.0
pandas_ta>=0.3.14b
numpy>=1.26.0
loguru>=0.7.2
python-dotenv>=1.0.1
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pytest>=7.4.0
